+++ src/MySQLReplication/Definitions/ConstEventsNames.php
@@ -17,4 +17,5 @@ class ConstEventsNames
     const UPDATE = 'update';
     const WRITE = 'write';
     const MARIADB_GTID = 'mariadb gtid';
+    const FORMAT_DESCRIPTION = 'format description';
 }