+++ src/MySQLReplication/Event/DTO/FormatDescriptionEventDTO.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace MySQLReplication\Event\DTO;
+
+use MySQLReplication\Definitions\ConstEventsNames;
+
+/**
+ * Class FormatDescriptionEventDTO
+ * @package MySQLReplication\Event\DTO
+ */
+class FormatDescriptionEventDTO extends EventDTO
+{
+    /**
+     * @var string
+     */
+    private $type = ConstEventsNames::FORMAT_DESCRIPTION;
+
+    /**
+     * @return string
+     */
+    public function getType()
+    {
+        return $this->type;
+    }
+
+    /**
+     * @return string
+     */
+    public function __toString()
+    {
+        return PHP_EOL .
+            '=== Event ' . $this->getType() . ' === ' . PHP_EOL .
+            'Date: ' . $this->eventInfo->getDateTime() . PHP_EOL .
+            'Log position: ' . $this->eventInfo->getPos() . PHP_EOL .
+            'Event size: ' . $this->eventInfo->getSize() . PHP_EOL;
+    }
+
+    /**
+     * Specify data which should be serialized to JSON
+     * @link http://php.net/manual/en/jsonserializable.jsonserialize.php
+     * @return mixed data which can be serialized by <b>json_encode</b>,
+     * which is a value of any type other than a resource.
+     * @since 5.4.0
+     */
+    public function jsonSerialize()
+    {
+        return get_object_vars($this);
+    }
+}