+++ tests/Unit/Gtid/GtidCollectionTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace MySQLReplication\Unit\Gtid;
+namespace MySQLReplication\Tests\Unit\Gtid;
 
 use MySQLReplication\Gtid\Gtid;
 use MySQLReplication\Gtid\GtidCollection;
-use MySQLReplication\Unit\BaseTest;
+use MySQLReplication\Tests\Unit\BaseTest;
 
 /**
  * Class GtidCollectionTest